+++ packages/playground/CHANGELOG.md
@@ -1,5 +1,11 @@
 # playground
 
+## 1.0.2
+
+### Patch Changes
+
+- multipleLineBraces 改为 multiLineBraces，新增类型导出
+
 ## 1.0.1
 
 ### Patch Changes